introduce helper functions. Preserves DOI search result even after ref type changes
